feat: add checking exist file before unlink

+++ Modules/Smartlegal/Http/Controllers/DocMandatoryController.php
@@ -392,7 +392,7 @@ class DocMandatoryController extends Controller
             if ($file) {
                 if($file->txtFilename != 'default.pdf') {
                     $destroy = public_path($file->txtPath);
-                    unlink($destroy);
+                    if (file_exists($destroy)) unlink($destroy);
                 }
             }
             foreach ($logs as $log) {
